Add DepthAI SDK structure
