+++ depthai_demo.py
@@ -4,14 +4,15 @@ from itertools import cycle
 from pathlib import Path
 import cv2
 import depthai as dai
-
-from depthai_helpers.managers import NNetManager, PreviewManager, FPSHandler, PipelineManager, Previews, EncodingManager
-from depthai_helpers.version_check import check_depthai_version
 import platform
 
 from depthai_helpers.arg_manager import parse_args
 from depthai_helpers.config_manager import ConfigManager
-from depthai_helpers.utils import frame_norm, to_planar, to_tensor_result, load_module
+from depthai_helpers.version_check import check_depthai_version
+from depthai_sdk.fps import FPSHandler
+from depthai_sdk.previews import Previews
+from depthai_sdk.managers import NNetManager, PreviewManager, PipelineManager, EncodingManager
+from depthai_sdk.utils import to_planar, load_module
 
 DISP_CONF_MIN = int(os.getenv("DISP_CONF_MIN", 0))
 DISP_CONF_MAX = int(os.getenv("DISP_CONF_MAX", 255))
