+++ depthai_sdk/README.md
@@ -1,0 +1,3 @@
+# DepthAI SDK
+
+WIP
